WEEK 2/Day 5/Lab 2/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.1
langchain-text-splitters>=0.3.0
numpy>=1.26.0
pypdf>=5.1.0
pandas>=2.2.0
matplotlib>=3.9.0
pytest>=8.3.0
